move draw_left_panel from app.rs into new file app_panels.rs
